ATLP-524. Added condition for tmp keytab deletion

+++ src/main/java/org/ohdsi/webapi/service/AbstractDaoService.java
@@ -12,6 +12,7 @@ import com.odysseusinc.datasourcemanager.krblogin.KerberosService;
 import com.odysseusinc.datasourcemanager.krblogin.KrbConfig;
 import com.odysseusinc.datasourcemanager.krblogin.RuntimeServiceMode;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.ohdsi.webapi.GenerationStatus;
@@ -168,7 +169,9 @@ public abstract class AbstractDaoService {
     }
     try {
       FileUtils.forceDelete(temporaryDir);
-      FileUtils.forceDelete(krbConfig.getComponents().getKeytabPath().toFile());
+      if (StringUtils.isNotBlank(krbConfig.getComponents().getKeytabPath().toString())){
+        FileUtils.forceDelete(krbConfig.getComponents().getKeytabPath().toFile());
+      }
     } catch (IOException e) {
       log.warn(e);
     }
